# internal/adapters/middleware/rest/swagger.yaml
basePath: /api/v1
definitions:
  rest.CredentialInfo:
    properties:
      expiry:
        type: string
      valid:
        type: boolean
    type: object
  rest.SessionInfo:
    properties:
      api_key:
        type: string
      credentials:
        $ref: '#/definitions/rest.CredentialInfo'
      provider_name:
        type: string
    type: object
host: localhost:9123
info:
  contact:
    url: https://github.com/tdrn-org/pim-mcp
  description: MCP server providing Agent access to PIM services.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: PIM MCP Server REST API
  version: "1.0"
paths:
  /api/v1/login:
    post:
      consumes:
      - application/json
      description: Initiate PIM provider login for the current user
      parameters:
      - description: login using api_key
        in: body
        name: api_key
        schema:
          type: string
      produces:
      - text/plain
      responses:
        "302":
          description: Found
          schema:
            type: string
        "500":
          description: server error
          schema:
            type: string
      summary: Initiate PIM provider login
  /api/v1/ping:
    get:
      description: Ping the server to check general health
      produces:
      - text/plain
      responses:
        "200":
          description: ok
          schema:
            type: string
        "500":
          description: server error
          schema:
            type: string
      summary: Ping the server
  /api/v1/session:
    delete:
      description: Delete the session for the current user
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "500":
          description: server error
          schema:
            type: string
      summary: Delete the user session
    get:
      description: Get the session for the current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.SessionInfo'
        "500":
          description: server error
          schema:
            type: string
      summary: Get the user session
swagger: "2.0"